+++ src/main/kotlin/com/jkjamies/cammp/feature/presentationgenerator/data/FlowStateHolderRepositoryImpl.kt
@@ -6,11 +6,11 @@ import com.jkjamies.cammp.feature.presentationgenerator.domain.repository.FileSy
 import com.jkjamies.cammp.feature.presentationgenerator.domain.repository.FlowStateHolderRepository
 import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.FileSpec
+import com.squareup.kotlinpoet.FunSpec
 import com.squareup.kotlinpoet.KModifier
-import com.squareup.kotlinpoet.ParameterizedTypeName.Companion.parameterizedBy
+import com.squareup.kotlinpoet.ParameterSpec
 import com.squareup.kotlinpoet.PropertySpec
 import com.squareup.kotlinpoet.TypeSpec
-import com.squareup.kotlinpoet.asClassName
 import java.nio.file.Path
 import kotlin.io.path.exists
 
@@ -33,26 +33,45 @@ class FlowStateHolderRepositoryImpl(
             return FileGenerationResult(target, GenerationStatus.SKIPPED, fileName)
         }
 
-        val mutableStateFlowClass = ClassName("kotlinx.coroutines.flow", "MutableStateFlow")
-        val stateFlowClass = ClassName("kotlinx.coroutines.flow", "StateFlow")
-        val unitClass = Unit::class.asClassName()
+        val composableAnnotation = ClassName("androidx.compose.runtime", "Composable")
+        val stableAnnotation = ClassName("androidx.compose.runtime", "Stable")
+        val rememberFunction = ClassName("androidx.compose.runtime", "remember")
+        val navHostControllerClass = ClassName("androidx.navigation", "NavHostController")
+        val rememberNavControllerFunction = ClassName("androidx.navigation.compose", "rememberNavController")
+        val flowStateClassName = ClassName(packageName, flowName)
+
+        val rememberFun = FunSpec.builder("remember$flowName")
+            .addModifiers(KModifier.INTERNAL)
+            .addAnnotation(composableAnnotation)
+            .addParameter(
+                ParameterSpec.builder("navController", navHostControllerClass)
+                    .defaultValue("%T()", rememberNavControllerFunction)
+                    .build()
+            )
+            .returns(flowStateClassName)
+            .beginControlFlow("return %T(navController)", rememberFunction)
+            .addStatement("%T(navController)", flowStateClassName)
+            .endControlFlow()
+            .build()
 
         val flowClass = TypeSpec.classBuilder(flowName)
             .addModifiers(KModifier.INTERNAL)
-            .addProperty(
-                PropertySpec.builder("_state", mutableStateFlowClass.parameterizedBy(unitClass))
-                    .addModifiers(KModifier.PRIVATE)
-                    .initializer("MutableStateFlow(Unit)")
+            .addAnnotation(stableAnnotation)
+            .primaryConstructor(
+                FunSpec.constructorBuilder()
+                    .addParameter("navController", navHostControllerClass)
                     .build()
             )
             .addProperty(
-                PropertySpec.builder("state", stateFlowClass.parameterizedBy(unitClass))
-                    .initializer("_state")
+                PropertySpec.builder("navController", navHostControllerClass)
+                    .addModifiers(KModifier.PRIVATE)
+                    .initializer("navController")
                     .build()
             )
             .build()
 
         val fileSpec = FileSpec.builder(packageName, flowName)
+            .addFunction(rememberFun)
             .addType(flowClass)
             .build()