+++ src/main/kotlin/com/jkjamies/cammp/feature/presentationgenerator/data/ScreenStateHolderRepositoryImpl.kt
@@ -4,7 +4,9 @@ import com.jkjamies.cammp.feature.presentationgenerator.domain.model.FileGenerat
 import com.jkjamies.cammp.feature.presentationgenerator.domain.model.GenerationStatus
 import com.jkjamies.cammp.feature.presentationgenerator.domain.repository.FileSystemRepository
 import com.jkjamies.cammp.feature.presentationgenerator.domain.repository.ScreenStateHolderRepository
+import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.FileSpec
+import com.squareup.kotlinpoet.FunSpec
 import com.squareup.kotlinpoet.KModifier
 import com.squareup.kotlinpoet.TypeSpec
 import java.nio.file.Path
@@ -30,12 +32,27 @@ class ScreenStateHolderRepositoryImpl(
             return FileGenerationResult(target, GenerationStatus.SKIPPED, fileName)
         }
 
+        val composableAnnotation = ClassName("androidx.compose.runtime", "Composable")
+        val stableAnnotation = ClassName("androidx.compose.runtime", "Stable")
+        val rememberFunction = ClassName("androidx.compose.runtime", "remember")
+        val stateHolderClassName = ClassName(packageName, className)
+
+        val rememberFun = FunSpec.builder("remember$className")
+            .addModifiers(KModifier.INTERNAL)
+            .addAnnotation(composableAnnotation)
+            .returns(stateHolderClassName)
+            .beginControlFlow("return %T()", rememberFunction)
+            .addStatement("%T()", stateHolderClassName)
+            .endControlFlow()
+            .build()
+
         val stateHolderClass = TypeSpec.classBuilder(className)
             .addModifiers(KModifier.INTERNAL)
-            .addKdoc("TODO: state for $screenName")
+            .addAnnotation(stableAnnotation)
             .build()
 
         val fileSpec = FileSpec.builder(packageName, className)
+            .addFunction(rememberFun)
             .addType(stateHolderClass)
             .build()